Move rust-common into labrinth
