+++ apps/labrinth/Cargo.toml
@@ -36,9 +36,10 @@ reqwest = { version = "0.11.18", features = ["json", "multipart"] }
 hyper = { version = "0.14", features = ["full"] }
 hyper-tls = "0.5.0"
 
+serde = { version = "1.0", features = ["derive"] }
+serde_bytes = "0.11"
 serde_json = "1.0"
 serde_cbor = "0.11"
-serde = { version = "1.0", features = ["derive"] }
 serde_with = "3.0.0"
 chrono = { version = "0.4.26", features = ["serde"] }
 yaserde = "0.12.0"
@@ -125,8 +126,6 @@ jemallocator = { version = "0.5.4", optional = true }
 async-stripe = { version = "0.39.1", features = ["runtime-tokio-hyper-rustls"] }
 rusty-money = "0.4.1"
 json-patch = "*"
-
-rust-common = { path = "../../packages/rust-common" }
 
 [dev-dependencies]
 actix-http = "3.4.0"
