+++ apps/labrinth/src/common/networking/message.rs
@@ -1,5 +1,5 @@
-use crate::ids::UserId;
-use crate::users::UserStatus;
+use crate::common::ids::UserId;
+use crate::common::users::UserStatus;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
