+++ apps/labrinth/src/common/networking/serialization.rs
