+++ apps/labrinth/src/database/models/ids.rs
@@ -1,9 +1,9 @@
 use super::DatabaseError;
+use crate::common::ids::base62_impl::to_base62;
+use crate::common::ids::{random_base62_rng, random_base62_rng_range};
 use censor::Censor;
 use rand::SeedableRng;
 use rand_chacha::ChaCha20Rng;
-use rust_common::ids::base62_impl::to_base62;
-use rust_common::ids::{random_base62_rng, random_base62_rng_range};
 use serde::{Deserialize, Serialize};
 use sqlx::sqlx_macros::Type;
