+++ apps/labrinth/src/database/models/pat_item.rs
@@ -1,11 +1,11 @@
 use super::ids::*;
+use crate::common::ids::base62_impl::parse_base62;
 use crate::database::models::DatabaseError;
 use crate::database::redis::RedisPool;
 use crate::models::pats::Scopes;
 use chrono::{DateTime, Utc};
 use dashmap::DashMap;
 use futures::TryStreamExt;
-use rust_common::ids::base62_impl::parse_base62;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Debug, Display};
 use std::hash::Hash;
