+++ apps/labrinth/src/database/models/session_item.rs
@@ -1,9 +1,9 @@
 use super::ids::*;
+use crate::common::ids::base62_impl::parse_base62;
 use crate::database::models::DatabaseError;
 use crate::database::redis::RedisPool;
 use chrono::{DateTime, Utc};
 use dashmap::DashMap;
-use rust_common::ids::base62_impl::parse_base62;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Debug, Display};
 use std::hash::Hash;
