+++ apps/labrinth/src/database/models/user_item.rs
@@ -1,12 +1,12 @@
 use super::ids::{ProjectId, UserId};
 use super::{CollectionId, ReportId, ThreadId};
+use crate::common::ids::base62_impl::{parse_base62, to_base62};
 use crate::database::models;
 use crate::database::models::{DatabaseError, OrganizationId};
 use crate::database::redis::RedisPool;
 use crate::models::users::Badges;
 use chrono::{DateTime, Utc};
 use dashmap::DashMap;
-use rust_common::ids::base62_impl::{parse_base62, to_base62};
 use serde::{Deserialize, Serialize};
 use std::fmt::{Debug, Display};
 use std::hash::Hash;
