+++ apps/labrinth/src/lib.rs
@@ -25,6 +25,8 @@ use crate::{
     util::env::{parse_strings_from_var, parse_var},
 };
 
+pub mod common;
+
 pub mod auth;
 pub mod clickhouse;
 pub mod database;
@@ -297,8 +299,10 @@ pub fn app_setup(
     }
 
     let ip_salt = Pepper {
-        pepper: rust_common::ids::Base62Id(rust_common::ids::random_base62(11))
-            .to_string(),
+        pepper: crate::common::ids::Base62Id(
+            crate::common::ids::random_base62(11),
+        )
+        .to_string(),
     };
 
     let payouts_queue = web::Data::new(PayoutsQueue::new());
