+++ apps/labrinth/src/routes/internal/admin.rs
@@ -74,7 +74,7 @@ pub async fn count_download(
     let project_id: crate::database::models::ids::ProjectId =
         download_body.project_id.into();
 
-    let id_option = rust_common::ids::base62_impl::parse_base62(
+    let id_option = crate::common::ids::base62_impl::parse_base62(
         &download_body.version_name,
     )
     .ok()
