+++ apps/labrinth/src/routes/internal/billing.rs
@@ -1,4 +1,5 @@
 use crate::auth::{get_user_from_headers, send_email};
+use crate::common::ids::base62_impl::{parse_base62, to_base62};
 use crate::database::models::charge_item::ChargeItem;
 use crate::database::models::{
     generate_charge_id, generate_user_subscription_id, product_item,
@@ -17,7 +18,6 @@ use crate::routes::ApiError;
 use actix_web::{delete, get, patch, post, web, HttpRequest, HttpResponse};
 use chrono::Utc;
 use log::{info, warn};
-use rust_common::ids::base62_impl::{parse_base62, to_base62};
 use rust_decimal::prelude::ToPrimitive;
 use rust_decimal::Decimal;
 use serde::Serialize;
