+++ apps/labrinth/src/routes/internal/flows.rs
@@ -1,6 +1,8 @@
 use crate::auth::email::send_email;
 use crate::auth::validate::get_user_record_from_bearer_token;
 use crate::auth::{get_user_from_headers, AuthProvider, AuthenticationError};
+use crate::common::ids::base62_impl::{parse_base62, to_base62};
+use crate::common::ids::random_base62_rng;
 use crate::database::models::flow_item::Flow;
 use crate::database::redis::RedisPool;
 use crate::file_hosting::FileHost;
@@ -23,8 +25,6 @@ use chrono::{Duration, Utc};
 use rand_chacha::rand_core::SeedableRng;
 use rand_chacha::ChaCha20Rng;
 use reqwest::header::AUTHORIZATION;
-use rust_common::ids::base62_impl::{parse_base62, to_base62};
-use rust_common::ids::random_base62_rng;
 use serde::{Deserialize, Serialize};
 use sqlx::postgres::PgPool;
 use std::collections::HashMap;
