+++ apps/labrinth/src/routes/mod.rs
@@ -117,7 +117,7 @@ pub enum ApiError {
     #[error("Captcha Error. Try resubmitting the form.")]
     Turnstile,
     #[error("Error while decoding Base62: {0}")]
-    Decoding(#[from] rust_common::ids::DecodingError),
+    Decoding(#[from] crate::common::ids::DecodingError),
     #[error("Image Parsing Error: {0}")]
     ImageParse(#[from] image::ImageError),
     #[error("Password Hashing Error: {0}")]
