+++ apps/labrinth/src/routes/v3/analytics_get.rs
@@ -1,4 +1,5 @@
 use super::ApiError;
+use crate::common::ids::base62_impl::to_base62;
 use crate::database;
 use crate::database::redis::RedisPool;
 use crate::models::teams::ProjectPermissions;
@@ -13,7 +14,6 @@ use crate::{
 };
 use actix_web::{web, HttpRequest, HttpResponse};
 use chrono::{DateTime, Duration, Utc};
-use rust_common::ids::base62_impl::to_base62;
 use serde::{Deserialize, Serialize};
 use sqlx::postgres::types::PgInterval;
 use sqlx::PgPool;
