+++ apps/labrinth/src/routes/v3/organizations.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 
 use super::ApiError;
 use crate::auth::{filter_visible_projects, get_user_from_headers};
+use crate::common::ids::base62_impl::parse_base62;
 use crate::database::models::team_item::TeamMember;
 use crate::database::models::{
     generate_organization_id, team_item, Organization,
@@ -21,7 +22,6 @@ use crate::util::validate::validation_errors_to_string;
 use crate::{database, models};
 use actix_web::{web, HttpRequest, HttpResponse};
 use futures::TryStreamExt;
-use rust_common::ids::base62_impl::parse_base62;
 use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use sqlx::PgPool;
