+++ apps/labrinth/src/routes/v3/projects.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 
 use crate::auth::checks::{filter_visible_versions, is_visible_project};
 use crate::auth::{filter_visible_projects, get_user_from_headers};
+use crate::common::ids::base62_impl::parse_base62;
 use crate::database::models::notification_item::NotificationBuilder;
 use crate::database::models::project_item::{GalleryItem, ModCategory};
 use crate::database::models::thread_item::ThreadMessageBuilder;
@@ -32,7 +33,6 @@ use actix_web::{web, HttpRequest, HttpResponse};
 use chrono::Utc;
 use futures::TryStreamExt;
 use itertools::Itertools;
-use rust_common::ids::base62_impl::parse_base62;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use sqlx::PgPool;
