+++ apps/labrinth/src/routes/v3/reports.rs
@@ -1,4 +1,5 @@
 use crate::auth::{check_is_moderator_from_headers, get_user_from_headers};
+use crate::common::ids::base62_impl::parse_base62;
 use crate::database;
 use crate::database::models::image_item;
 use crate::database::models::thread_item::{
@@ -17,7 +18,6 @@ use crate::util::img;
 use actix_web::{web, HttpRequest, HttpResponse};
 use chrono::Utc;
 use futures::StreamExt;
-use rust_common::ids::base62_impl::parse_base62;
 use serde::Deserialize;
 use sqlx::PgPool;
 use validator::Validate;
