+++ apps/labrinth/src/routes/v3/versions.rs
@@ -5,6 +5,7 @@ use crate::auth::checks::{
     filter_visible_versions, is_visible_project, is_visible_version,
 };
 use crate::auth::get_user_from_headers;
+use crate::common::ids::base62_impl::parse_base62;
 use crate::database;
 use crate::database::models::loader_fields::{
     self, LoaderField, LoaderFieldEnumValue, VersionField,
@@ -28,7 +29,6 @@ use crate::util::img;
 use crate::util::validate::validation_errors_to_string;
 use actix_web::{web, HttpRequest, HttpResponse};
 use itertools::Itertools;
-use rust_common::ids::base62_impl::parse_base62;
 use serde::{Deserialize, Serialize};
 use sqlx::PgPool;
 use validator::Validate;
