+++ apps/labrinth/src/search/indexing/mod.rs
@@ -1,6 +1,7 @@
 /// This module is used for the indexing from any source.
 pub mod local_import;
 
+use crate::common::ids::base62_impl::to_base62;
 use crate::database::redis::RedisPool;
 use crate::search::{SearchConfig, UploadSearchProject};
 use local_import::index_local;
@@ -8,7 +9,6 @@ use log::info;
 use meilisearch_sdk::client::{Client, SwapIndexes};
 use meilisearch_sdk::indexes::Index;
 use meilisearch_sdk::settings::{PaginationSetting, Settings};
-use rust_common::ids::base62_impl::to_base62;
 use sqlx::postgres::PgPool;
 use thiserror::Error;
 #[derive(Error, Debug)]
