+++ packages/app-lib/Cargo.toml
@@ -63,7 +63,7 @@ base64 = "0.22.0"
 
 sqlx = { version = "0.8.2", features = [ "runtime-tokio", "sqlite", "macros" ] }
 
-rust-common = { path = "../rust-common" }
+labrinth = { path = "../../apps/labrinth" }
 
 [target.'cfg(windows)'.dependencies]
 winreg = "0.52.0"
