--- packages/rust-common/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "@modrinth/networking",
-  "scripts": {
-    "build": "cargo build --release",
-    "lint": "cargo fmt --check && cargo clippy --all-targets -- -D warnings",
-    "fix": "cargo fmt && cargo clippy --fix",
-    "test": "cargo test"
-  }
-}
